move logic to ffs client
