+++ cli-docs/pow/pow_ffs_get.md
@@ -15,7 +15,7 @@ pow ffs get [cid] [output file path] [flags]
 ```
   -f, --folder                Indicates that the retrieved Cid is a folder
   -h, --help                  help for get
-      --ipfsrevproxy string   Powergate IPFS reverse proxy multiaddr (default "/ip4/127.0.0.1/tcp/6003")
+      --ipfsrevproxy string   Powergate IPFS reverse proxy multiaddr (default "127.0.0.1:6003")
   -t, --token string          token of the request
 ```
