+++ cmd/pow/cmd/ffs_stage.go
@@ -4,16 +4,11 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"net/http"
 	"os"
 	"time"
 
 	"github.com/caarlos0/spin"
 	"github.com/ipfs/go-cid"
-	files "github.com/ipfs/go-ipfs-files"
-	httpapi "github.com/ipfs/go-ipfs-http-client"
-	"github.com/ipfs/interface-go-ipfs-core/options"
-	"github.com/multiformats/go-multiaddr"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 	"github.com/textileio/powergate/util"
@@ -21,7 +16,7 @@ import (
 
 func init() {
 	ffsStageCmd.Flags().StringP("token", "t", "", "FFS access token")
-	ffsStageCmd.Flags().String("ipfsrevproxy", "/ip4/127.0.0.1/tcp/6003", "Powergate IPFS reverse proxy multiaddr")
+	ffsStageCmd.Flags().String("ipfsrevproxy", "127.0.0.1:6003", "Powergate IPFS reverse proxy multiaddr")
 
 	ffsCmd.AddCommand(ffsStageCmd)
 }
@@ -50,83 +45,22 @@ var ffsStageCmd = &cobra.Command{
 			Fatal(fmt.Errorf("getting file/folder information: %s", err))
 		}
 		var cid cid.Cid
+		s := spin.New("%s Staging specified asset in FFS hot storage...")
+		s.Start()
 		if fi.IsDir() {
-			token := viper.GetString("token")
-			cid, err = addFolder(args[0], token)
-			if err != nil {
-				Fatal(fmt.Errorf("staging folder: %s", err))
-			}
+			cid, err = fcClient.FFS.StageFolder(authCtx(ctx), viper.GetString("ipfsrevproxy"), args[0])
+			s.Stop()
+			checkErr(err)
 		} else {
 			f, err := os.Open(args[0])
 			checkErr(err)
 			defer func() { checkErr(f.Close()) }()
 
-			s := spin.New("%s Staging specified file in FFS hot storage...")
-			s.Start()
 			ptrCid, err := fcClient.FFS.Stage(authCtx(ctx), f)
-			s.Stop()
 			checkErr(err)
 			cid = *ptrCid
 		}
+		s.Stop()
 		Success("Staged asset in FFS hot storage with cid: %s", util.CidToString(cid))
 	},
-}
-
-func addFolder(folderPath string, ffsToken string) (cid.Cid, error) {
-	ipfsRevProxy := viper.GetString("ipfsrevproxy")
-	ipfs, err := newDecoratedIPFSAPI(ipfsRevProxy, ffsToken)
-	if err != nil {
-		return cid.Undef, fmt.Errorf("creating IPFS HTTP client: %s", err)
-	}
-
-	stat, err := os.Lstat(folderPath)
-	if err != nil {
-		return cid.Undef, err
-	}
-	ff, err := files.NewSerialFile(folderPath, false, stat)
-	if err != nil {
-		return cid.Undef, err
-	}
-	defer func() {
-		if err := ff.Close(); err != nil {
-			Fatal(fmt.Errorf("closing folder: %s", err))
-		}
-	}()
-	opts := []options.UnixfsAddOption{
-		options.Unixfs.CidVersion(1),
-		options.Unixfs.Pin(false),
-	}
-	pth, err := ipfs.Unixfs().Add(context.Background(), files.ToDir(ff), opts...)
-	if err != nil {
-		return cid.Undef, err
-	}
-
-	return pth.Cid(), nil
-}
-
-func newDecoratedIPFSAPI(ipfsRevProxyMaddr, ffsToken string) (*httpapi.HttpApi, error) {
-	ma, _ := multiaddr.NewMultiaddr(ipfsRevProxyMaddr)
-	customClient := http.DefaultClient
-	customClient.Transport = newFFSHeaderDecorator(ffsToken)
-	ipfs, err := httpapi.NewApiWithClient(ma, customClient)
-	if err != nil {
-		return nil, err
-	}
-	return ipfs, nil
-}
-
-type ffsHeaderDecorator struct {
-	ffsToken string
-}
-
-func newFFSHeaderDecorator(ffsToken string) *ffsHeaderDecorator {
-	return &ffsHeaderDecorator{
-		ffsToken: ffsToken,
-	}
-}
-
-func (fhd ffsHeaderDecorator) RoundTrip(req *http.Request) (*http.Response, error) {
-	req.Header["x-ipfs-ffs-auth"] = []string{fhd.ffsToken}
-
-	return http.DefaultTransport.RoundTrip(req)
 }
